requirements.txt
flask
sqlalchemy
flask-sqlalchemy
jinja2
python-dotenv